+++ smartsheet/models/create_report_result.py
@@ -1,0 +1,105 @@
+# pylint: disable=C0111,R0902,R0904,R0912,R0913,R0915,E1101
+# Smartsheet Python SDK.
+#
+# Copyright 2018 Smartsheet.com, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"): you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+
+from __future__ import absolute_import
+
+from typing import Optional
+
+from smartsheet.models.enums.access_level import AccessLevel
+from .report_column import ReportColumn
+
+from ..types import Boolean, EnumeratedValue, Number, String, TypedList, json
+from ..util import deserialize, serialize
+
+
+class CreateReportResult:
+    """Smartsheet CreateReportResult data model."""
+
+    def __init__(self, props=None, base_obj=None):
+        """Initialize the CreateReportResult model."""
+        self._base = None
+        if base_obj is not None:
+            self._base = base_obj
+
+        self._id = Number()
+        self._name = String()
+        self._access_level = EnumeratedValue(AccessLevel)
+        self._permalink = String()
+        self._is_summary_report = Boolean()
+        self._columns = TypedList(ReportColumn)
+
+        if props:
+            deserialize(self, props)
+
+        self.__initialized = True
+
+    @property
+    def id(self) -> Optional[int]:
+        return self._id.value
+
+    @id.setter
+    def id(self, value: int) -> None:
+        self._id.value = value
+
+    @property
+    def name(self) -> Optional[str]:
+        return self._name.value
+
+    @name.setter
+    def name(self, value: str) -> None:
+        self._name.value = value
+
+    @property
+    def access_level(self):
+        return self._access_level.value
+
+    @access_level.setter
+    def access_level(self, value) -> None:
+        self._access_level.set(value)
+
+    @property
+    def permalink(self) -> Optional[str]:
+        return self._permalink.value
+
+    @permalink.setter
+    def permalink(self, value: str) -> None:
+        self._permalink.value = value
+
+    @property
+    def is_summary_report(self) -> Optional[bool]:
+        return self._is_summary_report.value
+
+    @is_summary_report.setter
+    def is_summary_report(self, value: bool) -> None:
+        self._is_summary_report.value = value
+
+    @property
+    def columns(self) -> TypedList:
+        return self._columns
+
+    @columns.setter
+    def columns(self, value) -> None:
+        self._columns.load(value)
+
+    def to_dict(self):
+        return serialize(self)
+
+    def to_json(self):
+        return json.dumps(self.to_dict())
+
+    def __str__(self):
+        return self.to_json()