+++ smartsheet/models/enums/seat_type.py
@@ -5,11 +5,13 @@ class SeatType(str, Enum):
     GUEST = 'GUEST'
     MEMBER = 'MEMBER'
     PROVISIONAL_MEMBER = 'PROVISIONAL_MEMBER'
+    CONTRIBUTOR = 'CONTRIBUTOR'
 
 
 class DowngradeSeatType(str, Enum):
     VIEWER = 'VIEWER'
     GUEST = 'GUEST'
+    CONTRIBUTOR = 'CONTRIBUTOR'
 
 
 class UpgradeSeatType(str, Enum):