+++ smartsheet/models/report_scope_inclusion.py
@@ -1,0 +1,68 @@
+# pylint: disable=C0111,R0902,R0904,R0912,R0913,R0915,E1101
+# Smartsheet Python SDK.
+#
+# Copyright 2018 Smartsheet.com, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"): you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+
+from __future__ import absolute_import
+
+from smartsheet.models.enums.report_asset_type import ReportAssetType
+
+from ..types import EnumeratedValue, Number, json
+from ..util import deserialize, serialize
+
+class ReportScopeInclusion:
+
+    """Smartsheet ReportScopeInclusion data model."""
+
+    def __init__(self, props=None, base_obj=None):
+        """Initialize the ReportScopeInclusion model."""
+        self._base = None
+        if base_obj is not None:
+            self._base = base_obj
+
+        self._asset_type = EnumeratedValue(ReportAssetType)
+        self._asset_id = Number()
+
+        if props:
+            deserialize(self, props)
+
+    @property
+    def asset_id(self) -> int:
+        return self._asset_id.value
+
+    @asset_id.setter
+    def asset_id(self, value: int):
+        self._asset_id.value = value
+
+    @property
+    def asset_type(self) -> ReportAssetType:
+        return self._asset_type.value
+
+    @asset_type.setter
+    def asset_type(self, value):
+        # Convert string to uppercase to match enum member names
+        if isinstance(value, str):
+            self._asset_type.set(value.upper())
+        else:
+            self._asset_type.set(value)
+
+    def to_dict(self) -> dict:
+        return serialize(self)
+
+    def to_json(self) -> str:
+        return json.dumps(self.to_dict())
+
+    def __str__(self) -> str:
+        return self.to_json()